# data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI营销战略执行智能体",
    "description": "用AI帮小企业把营销战略翻译成每天的执行动作",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/agent/chat": {
      "post": {
        "tags": [
          "Agent 对话"
        ],
        "summary": "Agent Chat",
        "description": "与营销 Agent 对话一轮。",
        "operationId": "agent_chat_api_agent_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/history": {
      "get": {
        "tags": [
          "Agent 对话"
        ],
        "summary": "Agent History",
        "description": "读取某会话的对话历史。",
        "operationId": "agent_history_api_agent_history_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/chat/stream": {
      "post": {
        "tags": [
          "Agent 对话"
        ],
        "summary": "Agent Chat Stream",
        "description": "流式对话（SSE）。\n\n逐段返回 ``data: {json}\\n\\n`` 事件：intent / thinking / tool / done / error。\n前端用 ``wx.request({enableChunked:true})`` 解析分块，实时渲染 Agent 思考过程。",
        "operationId": "agent_chat_stream_api_agent_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "Login",
        "description": "微信小程序登录。\n\n流程：\n  1. 接收 wx.login() 返回的 code\n  2. 调用微信 jscode2session 获取 openid\n  3. 用 openid 生成 user_id\n  4. 签发标准 JWT（HS256 + 环境变量密钥）\n\nParameters\n----------\nreq : LoginRequest\n    登录请求体，包含微信小程序code。\n\nReturns\n-------\nAuthResponse\n    认证响应，data 包含 token（标准JWT）、user_id、openid。\n\nRaises\n------\nHTTPException\n    - 422: code 为空。\n    - 401: 微信 code 校验失败。\n    - 502: 微信服务不可用。",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "用户自然语言输入"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "会话ID；留空则用 用户ID:default",
            "default": ""
          },
          "business_id": {
            "type": "string",
            "title": "Business Id",
            "description": "可选企业ID；留空时复用会话已绑定企业",
            "default": ""
          },
          "files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Files",
            "description": "上传文件路径列表（复盘用）"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "AgentChatRequest",
        "description": "Agent 对话请求。"
      },
      "AgentHistoryResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "title": "AgentHistoryResponse",
        "description": "通用响应。"
      },
      "AuthResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "响应状态码，0表示成功",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息",
            "default": "ok"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data",
            "description": "响应数据载荷"
          }
        },
        "type": "object",
        "title": "AuthResponse",
        "description": "认证统一响应格式。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "微信小程序登录code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "LoginRequest",
        "description": "登录请求（微信小程序code）。\n\nParameters\n----------\ncode : str\n    微信小程序登录code，由 wx.login() 返回。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}